% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{causalSamp}
\alias{causalSamp}
\title{Sample from a causal model}
\usage{
causalSamp(
  n,
  formulas = list(list(z ~ 1), list(x ~ z), list(y ~ x), list(~1)),
  pars,
  family,
  link = NULL,
  dat = NULL,
  control = list(),
  seed
)
}
\arguments{
\item{n}{number of samples required}

\item{formulas}{list of lists of formulas}

\item{pars}{list of lists of parameters}

\item{family}{families for Z,X,Y and copula}

\item{link}{list of link functions (not currently used)}

\item{dat}{data frame of covariates}

\item{control}{options for the algorithm}

\item{seed}{random seed used for replication}
}
\value{
A data frame containing the simulated data.
}
\description{
Sample from a causal model
}
\details{
Samples from a given causal model using rejection sampling (or,
if everything is discrete, direct sampling).

The order of variables should be \code{z}, \code{x} then \code{y} (possibly
followed by \code{cop}).

We use the following codes for different families of distributions:
0 = binary
1 = normal
2 = t-distribution
3 = gamma

The family variables for the copula are also numeric and taken from
\code{VineCopula}.
Use, for example, 1 for Gaussian, 2 for t, 3 for Clayton, 4 for Gumbel,
5 for Frank, 6 for Joe and 11 for FGM copulas.

\code{pars} should be a named list containing: either entries \code{z},
\code{x}, \code{y} and \code{cop}, or variable names that correspond to the LHS of
formulae in \code{formulas}.  Each of these should themselves be a list
containing \code{beta} (a vector of regression parameters) and (possibly)
\code{phi}, a dispersion parameter.

Currently link functions are assumed to be the identity for a Gaussian or
t-distribution, the log-link for a Gamma distribution, and the logit link
for the Bernoulli distribution.

Control parameters are \code{oversamp} (default value 10),
\code{max_oversamp} (default value 1000), \code{warn} (which currently does
nothing), \code{max_wt} which is set to 1, and increases each time the function
is recalled.
}
\examples{
pars <- list(z=list(beta=0, phi=1),
             x=list(beta=c(0,0.5), phi=1),
             y=list(beta=c(0,0.5), phi=0.5),
             cop=list(beta=1))
causalSamp(100, pars = pars)

}
