% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE_checking.R
\name{causl_checking}
\alias{causl_checking}
\title{Checking the ATE given a causl model}
\usage{
causl_checking(
  n,
  formulas = forms,
  family = fams,
  pars = pars,
  link = NULL,
  estimand = "ate",
  method = "inversion",
  control = list()
)
}
\arguments{
\item{n}{number of samples required}

\item{formulas}{list of lists of formulas}

\item{family}{families for variables and copula}

\item{pars}{list of lists of parameters}

\item{link}{list of link functions}

\item{estimand}{either \code{"ate"} (the default), \code{"att"}, \code{"att"}, or \code{"ato"}}

\item{method}{either \code{"inversion"} (the default), \code{"inversion_mv"}, or \code{"rejection"}}

\item{control}{list of options for the algorithm}

\item{causl_model}{object of class \code{causl_model}}
}
\value{
a list containing correct and biased ATE with estimated se
using sandwich method
}
\description{
Checking the ATE given a causl model
}
