% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula_fams.R
\name{causl_copula}
\alias{causl_copula}
\alias{gaussian_causl_cop}
\title{Copula family functions}
\usage{
gaussian_causl_cop(link)
}
\arguments{
\item{link}{link function}
}
\description{
Copula family functions
}
\details{
These return a list that contains, for each valid family:
\itemize{
\itemize{
\item \code{name}: the name of the family;
\item \code{ddist}: function  to evaluate the density;
\item \code{rdist}: function to obtain samples;
\item \code{pars}: character vector of the parameter names;
\item \code{default}: list of default values;
\item \code{link}: the chosen link function.
}
}
}
\section{Functions}{
\itemize{
\item \code{gaussian_causl_cop()}: Gaussian copula family

}}
