% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causl_model.R
\name{causl_model}
\alias{causl_model}
\title{Define a \code{causl_model} object}
\usage{
causl_model(
  formulas,
  family,
  pars,
  link,
  dat = NULL,
  method = "inversion",
  kwd = "cop",
  control = list()
)
}
\arguments{
\item{formulas}{list of lists of formulas}

\item{family}{families for variables and copula}

\item{pars}{list of lists of parameters}

\item{link}{list of link functions}

\item{dat}{optional data frame of covariates}

\item{method}{either \code{"inversion"} (the default), \code{"inversion_mv"}, or \code{"rejection"}}

\item{kwd}{word used for copula formula and parameters}

\item{control}{list of options for the algorithm}
}
\description{
This defines a \code{causl_model} object, that can be used either for simulation
or inference.
}
\details{
The components \code{formulas} and \code{family} must both be specified, and have
matching lengths.  If \code{pars} is specified, then the model can be used for
simulation and inference, if not then only for inference.  \code{link} is optional,
and if not specified default links will be used.
}
