% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_causl.R
\name{check_rej}
\alias{check_rej}
\alias{check_pars}
\title{Check parameters for univariate families}
\usage{
check_rej(formulas, family, pars, dims, kwd)

check_pars(formulas, family, pars, dummy_dat, LHSs, kwd, dims)
}
\arguments{
\item{formulas}{list of lists of formulas}

\item{family}{families for variables and copula}

\item{pars}{list of lists of parameters}

\item{dims}{number of variables in each class}

\item{dummy_dat}{a dummy dataset, as generated by \code{gen_dummy_dat()}}

\item{LHSs}{left-hand sides from \code{formulas}}
}
\description{
Checks existence of beta vectors and then assesses appropriate length
}
\section{Functions}{
\itemize{
\item \code{check_rej()}: Checks for rejection sampling

}}
