% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulas.R
\name{copula_density}
\alias{copula_density}
\alias{dGaussCop}
\alias{dtCop}
\alias{dfgmCopula}
\title{Density of a multivariate copula}
\usage{
dGaussCop(x, Sigma, log = FALSE, use_cpp = TRUE, N)

dtCop(x, Sigma, df, log = FALSE, use_cpp = TRUE)

dfgmCopula(x, alpha)
}
\arguments{
\item{x}{samples on (0,1)}

\item{Sigma}{collection of matrices}

\item{log}{logical: return log=density?}

\item{use_cpp}{logical: use the C routine?}

\item{N}{optional integer for number of covariance matrices}

\item{df}{degrees of freedom}

\item{alpha}{parameter for copula}
}
\value{
numeric vector of densities
}
\description{
Density of a multivariate copula
}
\details{
Computes the density for data from a
Gaussian or t-copula.  Currently \code{use_cpp} only
works for \code{dGaussCop}.
}
\section{Functions}{
\itemize{
\item \code{dGaussCop()}: Gaussian copula

\item \code{dtCop()}: t-Copula density

\item \code{dfgmCopula()}: bivariate FGM copula

}}
