% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulas.R
\name{cVCopula}
\alias{cVCopula}
\title{Vectorized conditional copula function}
\usage{
cVCopula(U, copula, param, par2, inverse = FALSE)
}
\arguments{
\item{U}{matrix of quantiles}

\item{copula}{family of copula to use}

\item{param}{vector of parameters}

\item{par2}{Degrees of freedom for t-copula}

\item{inverse}{should inverse CDF be returned?}
}
\description{
Vectorized conditional copula function
}
\details{
Should have \code{nrow(U) = length(param)}.
}
