% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulas.R
\name{dfgmCopula}
\alias{dfgmCopula}
\title{Density of a bivariate FGM copula}
\usage{
dfgmCopula(u1, u2, alpha)
}
\arguments{
\item{u1, u2}{vector of probabilities}

\item{alpha}{parameter for copula}
}
\value{
numeric vector of densities
}
\description{
Density of a bivariate FGM copula
}
