% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulas.R
\name{dGaussDiscCop2}
\alias{dGaussDiscCop2}
\title{Density of a Mixed Copula}
\usage{
dGaussDiscCop2(x, m, Sigma, eta, log = FALSE, useC = TRUE)
}
\arguments{
\item{x}{matrix of samples on (0,1)}

\item{m}{number of discrete variables}

\item{Sigma}{collection of matrices}

\item{eta}{eta matrix}

\item{log}{logical: return log=density?}

\item{useC}{logical: use the C routine?}
}
\value{
numeric vector of densities
}
\description{
Density of a Mixed Copula
}
