% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_copulas.R
\name{emp_cdf}
\alias{emp_cdf}
\title{Empirical CDF}
\usage{
emp_cdf(xo, inv = FALSE, zero = "min")
}
\arguments{
\item{xo}{vector of observed values}

\item{inv}{logical: should inverse CDF be returned}

\item{zero}{value to use as inverse for zero}
}
\description{
Empirical CDF
}
\details{
\code{zero} can be \code{"min"} (for the minimum value of \code{xo}), \code{"m1"} (for
the minimum minus 1) or \code{"mInf"} (for \code{-Inf}).
}
