% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_copulas.R
\name{emp_cop}
\alias{emp_cop}
\title{Empirical copula}
\usage{
emp_cop(u, pts = u, smoothing = c("none", "checkerboard"))
}
\arguments{
\item{u}{matrix of integral probability transformed values}

\item{pts}{index of points at which to define copula}

\item{smoothing}{method of smoothing to use}
}
\description{
Empirical copula
}
