% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_power.R
\name{estimate_power}
\alias{estimate_power}
\title{Estimate the power under a causl_model and estimator}
\usage{
estimate_power(
  causl_model,
  n,
  estimator,
  monte_carlo_sims = 500,
  alpha_level = 0.05
)
}
\arguments{
\item{causl_model}{the specified causal model}

\item{n}{number of observations drawn from causal model}

\item{estimator}{method to estimate the coefficient and standard errors}

\item{monte_carlo_sims}{the number of simulations. Defaults to 500}

\item{alpha_level}{significance level one reject}
}
\description{
Estimate the power under a causl_model and estimator
}
