% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{ests_ses}
\alias{ests_ses}
\title{Extract parameter estimates and standard errors}
\usage{
ests_ses(fit, beta, merged_formula, kwd)
}
\arguments{
\item{fit}{output of \code{optim}}

\item{beta}{output of \code{initializeParams2}}

\item{merged_formula}{formula with all variables on RHS}

\item{kwd}{keyword for copula variable}
}
\description{
Extract parameter estimates and standard errors
}
