% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs2.R
\name{eta_matrix}
\alias{eta_matrix}
\alias{mu_matrix}
\title{Get means and linear forms for set of variables}
\usage{
eta_matrix(dat, pars, full_form, LHS)

mu_matrix(dat, pars, full_form, LHS, link)
}
\arguments{
\item{dat}{a data.frame containing all relevant variables}

\item{pars}{a list whose entries include every variable in \code{LHS}}

\item{full_form}{a merged formula whose left hand sides are in \code{LHS}}

\item{LHS}{a character vector of variables}

\item{link}{list of link functions, one for each entry of \code{LHS}}
}
\description{
Both \code{eta_matrix} and \code{mu_matrix} produce a matrix
with \code{nrows(dat)} rows and a column for each entry in \code{LHS}.  In
the case of \code{eta_matrix} the entry is the corresponding linear form
\eqn{X\beta} and for \code{mu_matrix} it is the transformed \eqn{g^{-1}(X\beta)}
that is also \eqn{E [Y | X=x]}
}
\section{Functions}{
\itemize{
\item \code{mu_matrix()}: transformed to mean scale

}}
