% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{family_checks}
\alias{family_checks}
\alias{is.causl_family}
\alias{is_categorical}
\alias{is_discrete}
\title{Check properties of family variables}
\usage{
is.causl_family(x)

is_categorical(x)

is_discrete(x)
}
\arguments{
\item{x}{a family, either numerical, a name, or a \code{causl_family} object}
}
\description{
Check properties of family variables
}
\details{
If the specific function does not represent any family then \code{NA} is
returned.

\code{is_categorical} returns a logical indicating if the object is the
input object represents a categorical or ordinal variable.

\code{is_discrete} returns a logical indicating if the object is the
input object represents a discrete random variable.
}
\section{Functions}{
\itemize{
\item \code{is.causl_family()}: Check if object has class \code{causl_family}

\code{is.causl_family} checks if object has \code{"causl_family"} in its class list.

\item \code{is_categorical()}: Check if family is categorical

\item \code{is_discrete()}: Check if family is discrete

}}
