% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{family_list}
\alias{family_list}
\title{Obtain list of family functions}
\usage{
family_list(family, func_return = get_family)
}
\arguments{
\item{family}{numeric or character vector of families}

\item{func_return}{function to apply to list of families}
}
\description{
Obtain list of family functions from numeric or character representation
}
\examples{
family_list(c(1,3,5))
family_list(c("t","binomial"))


}
