% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\docType{data}
\name{family_vals}
\alias{family_vals}
\alias{familyVals}
\alias{copula_vals}
\title{Numbers for parametric families}
\format{
\code{family_vals} is a \code{data.frame} with 9 rows and 2 columns

\code{familyVals} is the same object as \code{family_vals}

\code{copula_vals} is a \code{data.frame} with 7 rows and 2 columns
}
\usage{
family_vals

familyVals

copula_vals
}
\description{
Data frames containing
\itemize{
\item \code{val}: an integer
\item \code{family}: a vector giving the associated parametric family for that integer.
}
The integer \code{val} may be used in place of the name of the parametric family
when specifying the \code{family} object.
}
\section{Functions}{
\itemize{
\item \code{familyVals}: Old name

\item \code{copula_vals}: Values for copula families

}}
\keyword{datasets}
