% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\docType{data}
\name{familyVals}
\alias{familyVals}
\title{Numbers for parametric families}
\format{
An object of class \code{data.frame} with 7 rows and 2 columns.
}
\usage{
familyVals
}
\description{
A data frame containing
\itemize{
\item \code{val}: an integer
\item \code{family}: a vector giving the associated parametric family for that integer.
}
}
\keyword{datasets}
