% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_causl}
\alias{fit_causl}
\alias{fitCausal}
\title{Fit multivariate copula regression model}
\usage{
fit_causl(
  dat,
  formulas = list(y ~ x, z ~ 1, ~x),
  family = rep(1, length(formulas)),
  link,
  cop_pars,
  use_cpp = TRUE,
  control = list(),
  other_pars = list()
)

fitCausal(
  dat,
  formulas = list(y ~ x, z ~ 1, ~x),
  family = rep(1, length(formulas)),
  link,
  par2,
  sandwich = TRUE,
  use_cpp = TRUE,
  control = list()
)
}
\arguments{
\item{dat}{data frame of observations}

\item{formulas}{list of model formulae, for Y, for the Z variables, and
finally for the copula}

\item{family}{families for the Y and Z distributions, and the copula. Should
be the same length as \code{formulas}}

\item{link}{link functions for each variable}

\item{cop_pars}{additional parameters for copula if required}

\item{use_cpp}{logical: should C++ routines be used?}

\item{control}{list of parameters to be passed to \code{optim}}

\item{other_pars}{list of other parameters to use (e.g. degrees of freedom for a t-distribution)}

\item{par2}{former name for \code{cop_pars} argument}

\item{sandwich}{logical: should sandwich standard errors be returned?}
}
\value{
Returns a list of class \code{cop_fit}.
}
\description{
Fit multivariate copula regression model
}
\details{
\code{forms} is list of three or more formulae giving
predictors of y-margin, z-margin(s) and interaction
parameters.  Fit is by maximum likelihood.

\code{control} has the same arguments as the argument in \code{optim}, as well
as \code{sandwich}, a logical indicating if sandwich estimates of standard errors
should be computed, \code{newton}, a logical which controls whether Newton iterates should be
performed at the end, and \code{cop} which can edit the restricted variable name
for the left-hand side of formulae.
Useful for altering are \code{trace} (1 shows steps of optimization) and
\code{maxit} for the number of steps.

The list \code{other_pars} should be named with the relevant variables, and
each entry should be a named list containing the relevant parameters.

\strong{Warning} By default, none of the variables should be called \code{cop}, as
this is reserved for the copula.  The reserved word can be changed using
the argument \code{cop} within control.
}
\section{Functions}{
\itemize{
\item \code{fitCausal()}: old name

}}
