% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_optim.R
\name{fit_optim}
\alias{fit_optim}
\title{Fit using maximum likelihood estimation}
\usage{
fit_optim(
  dat,
  full_form,
  family,
  fam_cop,
  link,
  mm,
  cop_pars,
  LHSs,
  other_pars,
  control,
  use_cpp
)
}
\arguments{
\item{dat}{data frame of observations}

\item{full_form}{full formula with all variables on right-hand side}

\item{family}{families for variables and copula as above. Should be the
same length as \code{formulas}, and in same order}

\item{fam_cop}{the family of the copula}

\item{link}{link functions for each variable}

\item{mm}{model matrix}

\item{cop_pars}{additional parameters for copula if required}

\item{LHSs}{character string of left hand sides of formula}

\item{other_pars}{list of other parameters to use (e.g. degrees of freedom for a t-distribution)}

\item{control}{list of parameters to be passed to \code{optim}}

\item{use_cpp}{logical: should C++ routines be used?}
}
\description{
Fit using maximum likelihood estimation
}
