% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{formula_tools}
\alias{formula_tools}
\alias{lhs}
\alias{lhs<-}
\alias{rhs_vars}
\title{Tools for manipulating formulas}
\usage{
lhs(formulas, surv = FALSE)

lhs(formulas) <- value

rhs_vars(formulas)
}
\arguments{
\item{formulas}{list of formulae}

\item{surv}{logical indicating whether to treat as survey data}

\item{value}{character vector to assign}
}
\description{
Tools for manipulating formulas
}
\details{
\code{lhs} returns a character vector containing left-hand sides of a
list of formulae.  If \code{surv=TRUE} then two responses are returned in the
event of the left-hand side being a valid \code{Surv} object.  \verb{lhs<-} allows
one to assign the left-hand sides of variables in the obvious way.

\code{rhs_vars} extracts all the variables used on the right-hand sides of a list
of formulas.
}
\section{Functions}{
\itemize{
\item \code{lhs()}: Obtain left-hand sides from list of formulas

\item \code{lhs(formulas) <- value}: Assign left-hand sides to list of formulas

\item \code{rhs_vars()}: Extract variables from right-hand sides

}}
