% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_params.R
\name{gen_cop_pars}
\alias{gen_cop_pars}
\title{Function to generate random copula parameters for simulation}
\usage{
gen_cop_pars(formulas, data, range = c(-1, 1))
}
\arguments{
\item{formulas}{formulas as specified in \code{rfrugalParam}}

\item{data}{dataset to obtain parameterization for}

\item{range}{range of parameters to target}
}
\value{
A list suitable for the \code{cop} entry of the \code{pars} argument of
\code{rfrugalParam}
}
\description{
Attempts to ensure that values after passing through the
standard link function used for Gaussian copulas will have the specified
value.  For other copulas this will not target the correct range, but it
can still be used by considering how the relevant link functions work for
the Gaussian and other copula.
}
