% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_power.R
\name{get_causal_estimate}
\alias{get_causal_estimate}
\title{A helper function to calculate estimates from data. Make better to allow
for more estimators.}
\usage{
get_causal_estimate(data, estimator)
}
\arguments{
\item{data}{the data}

\item{estimator}{method to estimate the coefficient and standard errors}
}
\description{
A helper function to calculate estimates from data. Make better to allow
for more estimators.
}
