% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{get_family}
\alias{get_family}
\alias{gaussian_causl_fam}
\alias{t_causl_fam}
\alias{Gamma_causl_fam}
\alias{binomial_causl_fam}
\alias{beta_causl_fam}
\alias{categorical_causl_fam}
\alias{ordinal_causl_fam}
\title{Return causl_fam function from integer index}
\usage{
get_family(val)

gaussian_causl_fam(link)

t_causl_fam(link)

Gamma_causl_fam(link)

binomial_causl_fam(link)

beta_causl_fam(link)

categorical_causl_fam(link)

ordinal_causl_fam(link)
}
\arguments{
\item{val}{integer corresponding to distributional family}

\item{link}{link function}
}
\description{
Return causl_fam function from integer index
}
\details{
The functions \code{gaussian_causl_fam()} etc. represent the functions that are
returned by \code{get_family()}.

A few function of this form can be defined by the user, and it should return
the following:
\itemize{
\item \code{name}: the name of the relevant family;
\item \code{ddist}: a function returning the density of the distributions;
\item \code{qdist}: a function returning the quantiles from probabilities;
\item \code{rdist}: a function to sample values from the distribution;
\item \code{pdist}: a cumulative distribution function;
\item \code{pars}: a list of the names of the parameters used;
\item \code{default}: a function that returns a list of the default values for an
observation and each of the parameters;
\item \code{link}: the specified link function.
}

The function should also give the output the class \code{"causl_family"}, so that
it is interpreted appropriately.  Note that \code{ddist} should have a \code{log}
argument, to allow the log-likelihood to be evaluated.

The only parameterization of the categorical family currently implemented
is the multivariate logistic parameterization.  For a random variable \eqn{X}
with \eqn{K} states, dependence on a vector \eqn{\boldsymbol{Z}} uses:
\deqn{\log \dfrac{P(X=k)}{P(X=1)} = \beta_{k}^T \boldsymbol{Z},}
and the \eqn{\beta_k} vectors are stored as \eqn{(\beta_2,\dots,\beta_K)}.

The ordinal family is parameterized using a variation of the ordinal
logistic regression model.  This takes the logits of entries in the cumulative
distribution function and treats the covariates variables linearly on that
scale.  Suppose \eqn{\boldsymbol{Z}} is the vector of covariates and there are
\eqn{K} levels, then
\deqn{\log \dfrac{P(X \leq k)}{P(X > k)} = \beta_k^T \boldsymbol{Z}.}
As in the categorical case, the vectors \eqn{\beta_k} are stored as
\eqn{(\beta_1,\ldots,\beta_{K-1})}.
}
\section{Functions}{
\itemize{
\item \code{gaussian_causl_fam()}: Gaussian distribution family

\item \code{t_causl_fam()}: Student's t distribution family

\item \code{Gamma_causl_fam()}: Gamma distribution family

\item \code{binomial_causl_fam()}: binomial distribution family

\item \code{beta_causl_fam()}: beta distribution family

\item \code{categorical_causl_fam()}: multinomial/categorical distribution family

\item \code{ordinal_causl_fam()}: ordinal categorical distribution family

}}
\seealso{
\link{family_vals}
}
