% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_power.R
\name{get_ipw_estimate}
\alias{get_ipw_estimate}
\title{A helper function to calculate ipw weights from data.}
\usage{
get_ipw_estimate(data, ipw_info)
}
\arguments{
\item{data}{the data}

\item{ipw_info}{information about the propensity score methods}
}
\description{
A helper function to calculate ipw weights from data.
}
