% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fix.R
\name{get_max_weights}
\alias{get_max_weights}
\title{Get maximum weight for each segment of a distribution}
\usage{
get_max_weights(pars, forms_X, fam_X, qden, fam_Z, ranges, link, ...)
}
\arguments{
\item{pars}{list with all regression parameters}

\item{forms_X}{formulae for treatments}

\item{fam_X, fam_Z}{vector of families for treatments and covariates}

\item{qden}{density of proposals}

\item{ranges}{range of segments}

\item{link}{link functions for treatments}
}
\description{
Get maximum weight for each segment of a distribution
}
