% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fix2.R
\name{get_max_weights3}
\alias{get_max_weights3}
\title{Get maximum weight for each segment of a distribution}
\usage{
get_max_weights3(pars, forms_X, fam_X, qden, fam_Z, quantiles, link, ...)
}
\arguments{
\item{pars}{list with all regression parameters}

\item{forms_X}{formulae for treatments}

\item{fam_X, fam_Z}{vector of families for treatments and covariates}

\item{qden}{density of proposals}

\item{quantiles}{data.frame of integers indicating quantile of observation}

\item{link}{link functions for treatments}
}
\description{
Get maximum weight for each segment of a distribution
}
\details{
\code{quantiles} should have the names of the \code{Z} variables
as its names, and they should be in the same order as \code{fam_Z}.  Similarly,
\code{forms_X} and \code{fam_X} should correspond.

\bf{Warning:} this function currently assumes that variables in \code{Z} are
in a dependence related order.
}
