% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs2.R
\name{get_X_dens}
\alias{get_X_dens}
\title{Function to obtain density from several linear models}
\usage{
get_X_dens(n, fam, eta, link, pars, LHS, sim = TRUE)
}
\arguments{
\item{n}{number of samples}

\item{fam}{vector of families}

\item{eta}{matrix of covariates}

\item{link}{vector of link functions}

\item{pars}{list of parameters}

\item{LHS}{vector of names}

\item{sim}{logical: should data be generated, or just the density function returned?}
}
\description{
Function to obtain density from several linear models
}
