% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\name{get_X_density}
\alias{get_X_density}
\title{Get density of treatments}
\usage{
get_X_density(dat, eta, phi, qden, family, link, par2, log = FALSE)
}
\arguments{
\item{dat}{data frame of variables to change conditional distribution of}

\item{eta}{list (or matrix) of linear forms}

\item{phi}{vector of dispersion coefficients}

\item{qden}{functions for densities used to simulate variables}

\item{family}{vector of distribution families}

\item{link}{link functions for GLMs}

\item{par2}{vector of degrees of freedom}

\item{log}{logical: should log-density be returned?}
}
\value{
a numeric vector of weights
}
\description{
Get density of treatments
}
