% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{insert_lev}
\alias{insert_lev}
\alias{rmv_lev}
\title{Insert/remove extra list level for \code{causl_family} objects}
\usage{
insert_lev(x, target_class = "causl_family")

rmv_lev(x, target_class = "causl_family")
}
\arguments{
\item{x}{list to insert level into}

\item{target_class}{class to insert/remove level above}
}
\description{
Insert/remove extra list level for \code{causl_family} objects
}
\section{Functions}{
\itemize{
\item \code{rmv_lev()}: remove a level of listing

}}
