% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_power.R
\name{ipw_weights_from}
\alias{ipw_weights_from}
\title{Computes inverse probability weights (IPW) for each observation in the data.}
\usage{
ipw_weights_from(data, ipw_info)
}
\arguments{
\item{data}{A data.frame containing the variables required by the propensity score model.}

\item{ipw_info}{A list containing the propensity score model specification, including:
\itemize{
\item \code{formula}: a formula specifying the model for the treatment assignment.
\item \code{model}: a function to fit the propensity score model (e.g., \code{glm}).
\item \code{family}: the error distribution and link function to be used in the model.
}}
}
\value{
A numeric vector of fitted propensity scores (IPW weights) for each observation.
}
\description{
This function fits the specified propensity score model to the data and returns
the predicted probabilities (propensity scores) as weights for use in IPW estimation.
}
