% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{is_categorical}
\alias{is_categorical}
\title{Check if family is categorical}
\usage{
is_categorical(x)
}
\arguments{
\item{x}{a family, either numerical, a name, or a \code{causl_family} object}
}
\description{
Check if family is categorical
}
\details{
Returns a logical indicating if the object is the input object
represents a categorical or ordinal variable.  If it cannot represent a
family then \code{NA} is returned.
}
