% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{lhs}
\alias{lhs}
\alias{lhs<-}
\title{Get LHSs of list of formulae}
\usage{
lhs(formulas, surv = FALSE)

lhs(formulas) <- value
}
\arguments{
\item{formulas}{list of formulae}

\item{surv}{logical: should survival responses be returned as two elements?}

\item{value}{charachter vector to assign}
}
\description{
Get LHSs of list of formulae
}
\details{
Returns character vector containing left-hand sides of a list of
formulae.  If \code{surv=TRUE} then two responses are returned in the event of the
left-hand side being a valid \code{Surv} object.
}
\section{Functions}{
\itemize{
\item \code{lhs(formulas) <- value}: Assign LHSs to list of formulas

}}
