% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{link.causl_family}
\alias{link.causl_family}
\alias{link.causl_copula}
\title{Obtain link from a \code{causl_family} or \code{causl_copula} obect}
\usage{
\method{link}{causl_family}(x, ...)

\method{link}{causl_copula}(x, ...)
}
\arguments{
\item{x}{an object of class \code{causl_family} or \code{causl_copula}}
}
\description{
Obtain link from a \code{causl_family} or \code{causl_copula} obect
}
\section{Functions}{
\itemize{
\item \code{link(causl_copula)}: method for \code{causl_copula} object

}}
