% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\name{link_apply}
\alias{link_apply}
\title{Helper function to apply link function}
\usage{
link_apply(eta, link, family_nm)
}
\arguments{
\item{eta}{inner product of \code{X} and \code{beta}}

\item{link}{link function}

\item{family_nm}{family of distributions to use}
}
\value{
inner product \code{eta} transformed by inverse of link function
}
\description{
Helper function to apply link function
}
