% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R, R/sample_data_funcs.R
\name{link_setup}
\alias{link_setup}
\title{Set up link functions}
\usage{
link_setup(link, family, vars)

link_setup(link, family, vars)
}
\arguments{
\item{link}{the input given to causalSamp()}

\item{family}{the list of families for Z,X and Y variables}

\item{vars}{a list of vectors of variable names with the same structure as \code{family}}
}
\description{
Set up link functions

Set up link functions
}
\details{
Note that \code{beta} and \code{lognormal} are not exponential
dispersion families.  For the \code{beta} distribution we use the approach
outlined in Cribari-Neta and Zeileis (2010), and for the \code{lognormal}
we just treat as a normal distribution that is then exponentiated.
}
\references{
F. Cribari-Neto, and A. Zeileis. Beta regression in R. \emph{Journal of Statistical Software} 34: 1-24, 2010.
}
