% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\docType{data}
\name{links_list}
\alias{links_list}
\alias{linksList}
\title{List of links available for each parametric family}
\format{
An object of class \code{list} of length 8.

\code{linksList} is the old name for \code{links_list}
}
\usage{
links_list

linksList
}
\description{
This is a named list whose entries are character vectors of allowed link
functions.
}
\section{Functions}{
\itemize{
\item \code{linksList}: Old name

}}
\keyword{datasets}
