% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\docType{data}
\name{linksList}
\alias{linksList}
\title{List of links available for each parametric family}
\format{
An object of class \code{list} of length 6.
}
\usage{
linksList
}
\description{
This is a named list whose entries are character vectors of allowed link
functions.
}
\keyword{datasets}
