% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{ll_frugal}
\alias{ll_frugal}
\title{Log-likelihood for frugal parameterization}
\usage{
ll_frugal(pars, dat, formulas, family, link, kwd = "cop")
}
\arguments{
\item{pars}{parameter values}

\item{dat}{\code{data.frame} containing data}

\item{formulas}{list of lists of formulas}

\item{family}{families for variables and copula}

\item{link}{list of link functions}

\item{kwd}{string to use for copula}
}
\description{
Log-likelihood for frugal parameterization
}
