% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_fix2.R
\name{match_rows}
\alias{match_rows}
\title{Match rows in a matrix with duplicates to set of unique values}
\usage{
match_rows(x, y, nomatch = NA_integer_)
}
\arguments{
\item{x}{matrix with unique rows}

\item{y}{matrix to be matched}

\item{nomatch}{value to insert when there is no match}
}
\description{
Match rows in a matrix with duplicates to set of unique values
}
