% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{merge_formulas}
\alias{merge_formulas}
\title{Combine multiple formulas}
\usage{
merge_formulas(formulas)
}
\description{
Take collection of formulae and create one formula with all variables on the
right-hand side of any of the originals.
}
