% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causl_model.R
\name{modify}
\alias{modify}
\alias{modify.default}
\title{Generic for modify}
\usage{
modify(x, ...)

\method{modify}{default}(x, ...)
}
\arguments{
\item{x}{object to be modified}

\item{...}{other arguments}
}
\description{
Generic for modify
}
\section{Methods (by class)}{
\itemize{
\item \code{modify(default)}: default method

}}
\seealso{
\code{\link[=modify.causl_model]{modify.causl_model()}}
}
