% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_family.R
\name{new_family}
\alias{new_family}
\title{Add a new parametric family}
\usage{
new_family(
  name,
  ddist,
  qdist,
  rdist,
  pdist,
  pars = c(x = 0, mu = 0, phi = 1, par2 = 1),
  links
)
}
\arguments{
\item{name}{string giving name for family}

\item{pars}{named vector of parameters used to evaluate function}

\item{links}{character vector of link functions to use with family}

\item{dens, quan, sim}{density, quantile and simulation functions (see details)}

\item{number}{optional number to specify}
}
\description{
Include a new parametric family by specifying its density, quantiles and a function to simulate with
}
