% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{nll}
\alias{nll}
\title{Negative log-likelihood of multivariate conditional copula}
\usage{
nll(beta, dat, mms, fam_cop = 1, fam = rep(1, nc), par2 = NULL, useC = TRUE)
}
\arguments{
\item{beta}{vector of parameters}

\item{dat}{output variables}

\item{mms}{list of model matrices and indices created by \code{fitCausal}}

\item{fam_cop}{copula family}

\item{fam}{univariate families for distributions}

\item{par2}{second parameter for t-distributions}

\item{useC}{logical: should C++ routine be used?}
}
\value{
Numeric value of the negative log-likelihood.
}
\description{
Negative log-likelihood of multivariate conditional copula
}
\details{
\code{fam_cop} has: 1=Gaussian, 2=t, 3:6 Archimidean, 11 FGM, and
 \code{fam} has 1=Gaussian, 2=t, 3=Gamma (with log-link), 5=truncated normal.
}
