% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{nll2}
\alias{nll2}
\title{Negative log-likelihood}
\usage{
nll2(
  theta,
  dat,
  mm,
  beta,
  phi,
  inCop,
  fam_cop = 1,
  family,
  link,
  cop_pars = NULL,
  useC = TRUE
)
}
\arguments{
\item{theta}{concatenated vector of parameters (\code{beta} followed by \code{phi})}

\item{dat}{matrix of data}

\item{mm}{model matrix for use with \code{beta}}

\item{beta}{(sparse) matrix of regression parameters for each variable and copula}

\item{phi}{vector of dispersion parameters}

\item{inCop}{vector of integers giving variables in \code{dat} to be included in copula}

\item{fam_cop, family}{integer and integer vector for copula and distribution families respectively}

\item{link}{vector of link functions}

\item{cop_pars}{other parameters for copula}

\item{useC}{logical: should Rcpp functions be used?}
}
\description{
Negative log-likelihood
}
\details{
The number of columns of \code{beta} should be the number of columns
in \code{dat} plus the number required to parameterize the copula.  The first
few columns and the entries in \code{phi} are assumed to be in the order of
those in \code{dat}.  If the \eqn{i}th
family for a variable does not require a dispersion parameter then the value of
\code{phi[i]} is ignored.
}
