% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{pair_copula_setup}
\alias{pair_copula_setup}
\title{Sets up copula quantities only}
\usage{
pair_copula_setup(formulas, family, pars, LHSs, quans, ord)
}
\arguments{
\item{formulas}{list of formulas for copula only}

\item{family}{list of families for copula only}

\item{pars}{list of copula parameters}

\item{LHSs}{left-hand sides for all variables}

\item{quans}{character vector of already existing variables to include}

\item{ord}{topological ordering}
}
\description{
Sets up copula quantities only
}
