% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intialize.R
\name{par_masks}
\alias{par_masks}
\title{Get parameter masks for regression parameters}
\usage{
par_masks(formulas, family = rep(1, nv), full_form)

par_masks(formulas, family = rep(1, nv), full_form)
}
\arguments{
\item{formulas}{formulas to create mask for}

\item{family}{vector or list of families}

\item{full_form}{(optionally) merged list of \code{formulas}}
}
\description{
Get parameter masks for regression parameters

Get parameter masks for regression parameters
}
