% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causl_model.R
\name{print.causl_model}
\alias{print.causl_model}
\title{Display output from \code{causl_model}}
\usage{
\method{print}{causl_model}(x, ...)
}
\arguments{
\item{x}{an object of class \code{causl_model}}

\item{...}{additional arguments (not used)}
}
\description{
Display output from \code{causl_model}
}
