% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{process_family}
\alias{process_family}
\title{Process input for family variables}
\usage{
process_family(family)
}
\arguments{
\item{family}{list of family parameters}
}
\description{
Process input for family variables
}
