% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_link.R
\name{process_family_link}
\alias{process_family_link}
\title{New function to sort out links and families simultaneously}
\usage{
process_family_link(family, link, dims, func_return = get_family)
}
\arguments{
\item{family}{families for variables and copula}

\item{link}{list of link functions}

\item{dims}{number of variables in each class}

\item{func_return}{function to use to process character arguments}
}
\description{
New function to sort out links and families simultaneously
}
