% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{process_inputs}
\alias{process_inputs}
\title{Process formulas, families and parameters}
\usage{
process_inputs(formulas, pars, family, link, kwd)
}
\arguments{
\item{formulas}{list of lists of formulas}

\item{pars}{list of lists of parameters}

\item{family}{families for Z,X,Y and copula}

\item{link}{list of link functions}

\item{kwd}{keyword for copula}
}
\description{
Process formulas, families and parameters
}
