% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_causl.R
\name{process_prespecified}
\alias{process_prespecified}
\title{Obtain quantiles for prespecified variables}
\usage{
process_prespecified(dat, prespec)
}
\arguments{
\item{dat}{data frame containing variables}

\item{prespec}{character vector of prespecified variables in \code{dat}}
}
\description{
Obtain quantiles for prespecified variables
}
\details{
Currently takes the rank of each entry, and subtracts 1/2 and
normalizes by the number of entries.  If there are \eqn{k} ties they are
randomly sorted with a uniform random variable
in the symmetric interval around the rank of width \eqn{k/n}.
}
