% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\name{rejectionWeights}
\alias{rejectionWeights}
\title{Get weights for rejection sampling}
\usage{
rejectionWeights(dat, mms, family, pars, qden)
}
\arguments{
\item{dat}{data frame of variables to change conditional distribution of}

\item{mms}{list of model matrices}

\item{family}{vector of distribution families}

\item{pars}{parameters for new distributions}

\item{qden}{functions for densities used to simulate variables}
}
\value{
a numeric vector of weights
}
\description{
Get weights for rejection sampling
}
