% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\name{rescale_cop}
\alias{rescale_cop}
\alias{rescaleCop}
\title{Rescale quantiles to conditional copula}
\usage{
rescale_cop(U, X, beta, family = 1, par2)

rescaleCop(U, X, beta, family = 1, par2)
}
\arguments{
\item{U}{matrix of quantiles}

\item{X}{model matrix of covariates}

\item{beta}{list of parameters (see details)}

\item{family}{variety of copula to use}

\item{par2}{additional parameter for some copulas}
}
\value{
vector of rescaled quantiles
}
\description{
Rescale quantiles to conditional copula
}
\details{
The variable to be transformed must be in the final column of
\code{U}, with variables being conditioned upon in the earlier columns.

\code{family} can be 1 for Gaussian, 2 for t, 3 for Clayton, 4 for
Gumbel, 5 for Frank, 6 for Joe and 11 for FGM copulas. Gamma distributed,
beta distributed or discrete respectively. \code{pars} should be a list
with entries \code{beta} and \code{phi}, as well as possibly \code{par2} if
\code{family=2}.
\code{U} should have the same length as \code{X} has rows, and \code{X}
should have the same number of columns as the length of \code{pars$beta}.
}
\section{Functions}{
\itemize{
\item \code{rescaleCop()}: Old name, now deprecated

}}
