% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\name{rescaleCop}
\alias{rescaleCop}
\title{Rescale quantiles to conditional copula}
\usage{
rescaleCop(U, X, pars, family = 1)
}
\arguments{
\item{U}{vector of quantiles}

\item{X}{model matrix of covariates}

\item{pars}{list of parameters (see details)}

\item{family}{variety of copula to use}
}
\value{
vector of rescaled quantiles
}
\description{
Rescale quantiles to conditional copula
}
\details{
\code{family} can be 1 for Gaussian, 2 for t, 3 for Clayton, 4 for
Gumbel, 5 for Frank, 6 for Joe and 11 for FGM copulas. Gamma distributed,
beta distributed or discrete respectively. \code{pars} should be a list
with entries \code{beta} and \code{phi}, as well as possibly \code{par2} if
\code{family=2}.
\code{U} should have the same length as \code{X} has rows, and \code{X}
should have the same number of columns as the length of \code{pars$beta}.
}
