% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\name{rescaleVar}
\alias{rescaleVar}
\title{Rescale quantiles to arbitrary random variable.}
\usage{
rescaleVar(U, X, pars, family = 1, link)
}
\arguments{
\item{U}{vector of quantiles}

\item{X}{model matrix of covariates}

\item{pars}{list of parameters (see details)}

\item{family}{family of distributions to use}

\item{link}{link function}
}
\value{
vector of rescaled variables
}
\description{
Rescale quantiles to arbitrary random variable.
}
\details{
\code{family} can be 1, 2, 3, 4 or 5 for Gaussian, t-distributed,
Gamma distributed, beta distributed or discrete respectively.
\code{pars} should be a list with entries \code{beta} and
\code{phi}, as well as possibly \code{par2} and \code{trunc} if the family
is set to 2 or 5.
\code{U} should have the same length as \code{X} has rows, and
\code{X} should have the same number of columns as the length of
\code{pars$beta}.
}
