% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_new.R
\name{rfrugal}
\alias{rfrugal}
\alias{rfrugalParam}
\title{Sample from a causal model}
\usage{
rfrugal(n, causl_model, control = list())

rfrugalParam(
  n,
  formulas = list(list(z ~ 1), list(x ~ z), list(y ~ x), list(~1)),
  family = c(1, 1, 1, 1),
  pars,
  link = NULL,
  dat = NULL,
  method = "inversion",
  control = list(),
  ...
)
}
\arguments{
\item{n}{number of samples required}

\item{causl_model}{object of class \code{causl_model}}

\item{control}{list of options for the algorithm}

\item{formulas}{list of lists of formulas}

\item{family}{families for variables and copula}

\item{pars}{list of lists of parameters}

\item{link}{list of link functions}

\item{dat}{optional data frame of covariates}

\item{method}{either \code{"inversion"} (the default), \code{"inversion_mv"}, or \code{"rejection"}}

\item{...}{other arguments, such as custom families}

\item{estimand}{quantity to control, default is \code{"ATE"}}
}
\value{
A data frame containing the simulated data.
}
\description{
Obtain samples from a causal model parameterized as in Evans and Didelez (2024).
}
\details{
Samples from a given causal model under the frugal
parameterization.

The entries for  \code{formula} and \code{family} should each be a
list with four entries, corresponding to the \eqn{Z}, \eqn{X}, \eqn{Y} and
the copula.  \code{formula} determines the model, so it is crucial that
every variable to be simulated is represented there exactly once.  Each
entry of that list can either be a single formula, or a list of formulae.
Each corresponding entry in \code{family} should be the same length as the
list in \code{formula} or of length 1 (in which case it will be repeated
for all the variables therein).

We use the following codes for different families of distributions:\tabular{rl}{
   val \tab family \cr
   0 \tab binomial \cr
   1 \tab gaussian \cr
   2 \tab t \cr
   3 \tab Gamma \cr
   4 \tab beta \cr
   5 \tab binomial \cr
   6 \tab lognormal \cr
   11 \tab ordinal \cr
   10 \tab categorical \cr
}


The family variables for the copula are also numeric and taken from
\code{VineCopula}. Use, for example, 1 for Gaussian, 2 for t, 3 for Clayton,
4 for Gumbel, 5 for Frank, 6 for Joe and 11 for FGM copulas.

\code{pars} should be a named list containing variable names that correspond
to the LHS of
formulae in \code{formulas}.  Each of these should themselves be a list
containing \code{beta} (a vector of regression parameters) and (possibly)
\code{phi}, a dispersion parameter.  For any discrete variable that is a
treatment, you can also specify \code{p}, an initial proportion to simulate
from (otherwise this defaults to 0.5).

Link functions for the Gaussian, t and Gamma distributions can be the
identity, inverse or log functions.  Gaussian and t-distributions default to
the identity, and Gamma to the log link.  For the Bernoulli the logit,
probit, and log links are available.

A variety of sampling methods are implemented.  The
inversion method with pair-copulas is the default (\code{method="inversion"}),
but we cam also use a multivariate copula (\code{method="inversion_mv"}) or even
rejection sampling (\code{method="rejection"}).

The only control parameters are \code{cop}: which gives a keyword for the
copula that defaults to \code{"cop"}; \code{quiet} which defaults to \code{FALSE} but will
reduce output if set to \code{TRUE}; and (if rejection sampling is selected)
\code{careful}: this logical enables one to implement the full rejection
sampling method, which means we do get exact samples (note this method
is generally very slow, especially if we have an outlying value, so the
default is \code{FALSE}).
}
\section{Functions}{
\itemize{
\item \code{rfrugalParam()}: old function for simulation

}}
\examples{
pars <- list(z=list(beta=0, phi=1),
             x=list(beta=c(0,0.5), phi=1),
             y=list(beta=c(0,0.5), phi=0.5),
             cop=list(beta=1))
rfrugalParam(100, pars = pars)


}
