% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_new.R
\name{rfrugalParam}
\alias{rfrugalParam}
\title{Sample from a causal model}
\usage{
rfrugalParam(
  n,
  formulas = list(list(z ~ 1), list(x ~ z), list(y ~ x), list(~1)),
  pars,
  family,
  link = NULL,
  dat = NULL,
  method = "rejection",
  control = list(),
  seed
)
}
\arguments{
\item{n}{number of samples required}

\item{formulas}{list of lists of formulas}

\item{pars}{list of lists of parameters}

\item{family}{families for Z,X,Y and copula}

\item{link}{list of link functions}

\item{dat}{data frame of covariates}

\item{method}{only \code{"rejection"} is valid}

\item{control}{list of options for the algorithm}

\item{seed}{random seed used for replication}

\item{data}{optional data frame of covariates}
}
\value{
A data frame containing the simulated data.
}
\description{
Obtain samples from a causal model using the rejection sampling approach of
Evans and Didelez (2021).
}
\details{
Samples from a given causal model using rejection sampling (or,
if everything is discrete, direct sampling).

The entries for  \code{formula} and \code{family} should each be a
list with four entries, corresponding to the \eqn{Z}, \eqn{X}, \eqn{Y} and
the copula.  \code{formula} determines the model, so it is crucial that
every variable to be simulated is represented there exactly once.  Each
entry of that list can either be a single formula, or a list of formulae.
Each corresponding entry in \code{family} should be the same length as the
list in \code{formula} or of length 1 (in which case it will be repeated
for all the variables therein).

We use the following codes for different families of distributions:
0 or 5 = binary;
1 = normal;
2 = t-distribution;
3 = gamma;
4 = beta;
6 = log-normal.

The family variables for the copula are also numeric and taken from
\code{VineCopula}.
Use, for example, 1 for Gaussian, 2 for t, 3 for Clayton, 4 for Gumbel,
5 for Frank, 6 for Joe and 11 for FGM copulas.

\code{pars} should be a named list containing: either entries \code{z},
\code{x}, \code{y} and \code{cop}, or variable names that correspond to the LHS of
formulae in \code{formulas}.  Each of these should themselves be a list
containing \code{beta} (a vector of regression parameters) and (possibly)
\code{phi}, a dispersion parameter.  For any discrete variable that is a
treatment, you can also specify \code{p}, an initial proportion to simulate
from (otherwise this defaults to 0.5).

Link functions for the Gaussian, t and Gamma distributions can be the
identity, inverse or log functions.  Gaussian and t-distributions default to
the identity, and Gamma to the log link.  For the Bernoulli the logit and
probit links are available.

Control parameters are \code{trace} (default value 0, increasing to 1
increases verbosity of output), \code{warn} (which currently does nothing),
\code{max_wt} which is set to 1, and increases each time the function
is recalled.
Control parameters also include \code{cop}, which gives a keyword for the
copula that defaults to \code{"cop"}.
}
\examples{
pars <- list(z=list(beta=0, phi=1),
             x=list(beta=c(0,0.5), phi=1),
             y=list(beta=c(0,0.5), phi=0.5),
             cop=list(beta=1))
causalSamp(100, pars = pars)


}
