% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulas.R
\name{rGaussCop}
\alias{rGaussCop}
\alias{rtCop}
\alias{rfgmCopula}
\title{Sample from Gaussian or t-copula}
\usage{
rGaussCop(n, Sigma)

rtCop(n, Sigma, df)

rfgmCopula(n, d = 2, alpha)
}
\arguments{
\item{n}{sample size}

\item{Sigma}{in which each slice is a correlation matrix}

\item{df}{degrees of freedom}

\item{d}{dimension of copula}

\item{alpha}{(vector of) parameter values}
}
\value{
A vector of the simulated random variables.
}
\description{
Sample from Gaussian or t-copula
}
\details{
Quicker than rCopula.

Note that \code{rfgmCopula} only works for \eqn{d = 2}.
}
\section{Functions}{
\itemize{
\item \code{rtCop()}: t-copula

\item \code{rfgmCopula()}: FGM-copula

}}
