% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{sandwich_errors}
\alias{sandwich_errors}
\title{Function to compute sandwich errors}
\usage{
sandwich_errors(func, pars, method = "simple", trace = FALSE, ...)
}
\arguments{
\item{func}{function for computation}

\item{pars}{parameter values at optimum}

\item{method}{method for \code{grad} to use}

\item{trace}{logical: should comments be shown}

\item{...}{other arguments used in \code{func}, including \code{dat} and \code{mm}}
}
\description{
Function to compute sandwich errors
}
