% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs2.R
\name{sim}
\alias{sim}
\title{Simulate initial X values}
\usage{
sim(n, fam, theta, offset, sim = TRUE)
}
\arguments{
\item{n}{number of observations}

\item{fam}{number for distribution family}

\item{theta}{parameters for model}

\item{offset}{optional mean correction}

\item{sim}{logical: should values be simulated?}
}
\value{
A list with two entries: \code{x} a vector of the simulated
values, and \code{qden}, which contains a function that evaluates to the
density of the distribution used to generate those values.
}
\description{
Simulate initial X values
}
\details{
Returns a list that includes a data frame containing a column
\code{x}, as well as the density that was used to generate it.  Possible
families are Gaussian (=1), t (=2), Exponential (=3), beta (=4)
Bernoulli/categorical (=5) and log-normal (=6).

For the exponential distribution, \code{theta} is the mean.
Beta can take one or two parameters, and if there is only
one it is just repeated.

The \code{offset} parameter alters the median for the normal and t-distributions,
or the median of the logarithm in the case of a log-normal.
}
