% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\name{sim_copula}
\alias{sim_copula}
\alias{sim_CopVal}
\title{Simulate copula values}
\usage{
sim_copula(dat, family, par, df, model_matrix)

sim_CopVal(dat, family, par, df, model_matrix)
}
\arguments{
\item{dat}{data frame with empty columns}

\item{family}{numeric indicator of copula type}

\item{par}{mandatory parameters}

\item{df}{optional parameters}

\item{model_matrix}{design matrix for covariates}
}
\value{
A data frame of the same dimension as \code{dat} containing the
simulated values.
}
\description{
Simulate copula values
}
\details{
Returns data frame containing columns \code{y}
and \code{z1, ..., zk}.

The family variables are numeric and taken from \code{VineCopula}.
Use, for example, 1 for Gaussian, 2 for t, 3 for Clayton, 4 for Gumbel,
5 for Frank, 6 for Joe and 11 for FGM copulas.
}
\section{Functions}{
\itemize{
\item \code{sim_CopVal()}: Old name, now deprecated

}}
