% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_multi.R
\name{sim_multi}
\alias{sim_multi}
\title{Simulate for single time-step}
\usage{
sim_multi(out, proc_inputs)
}
\arguments{
\item{out}{data frame for output}

\item{proc_inputs}{output of \code{process_inputs()}}

\item{control}{list of control parameters}
}
\description{
Simulate for single time-step
}
\details{
\code{sim_inversion} and \code{sim_rejection} correspond to
performing the sampling by inversion or using rejection sampling.\code{sim_multi} first
simulates from the copula then transforms to the correct margins in the correct causal ordering
}
