% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_survival.R
\name{sim_survival}
\alias{sim_survival}
\title{Simulate from a discrete time survival model}
\usage{
sim_survival(
  n,
  T = 1,
  formulas = list(list(z ~ 1), list(x ~ z), list(y ~ x), list(~1)),
  pars,
  family,
  lags,
  dat = NULL,
  control = list()
)
}
\arguments{
\item{n}{number of samples}

\item{T}{number of timepoints}

\item{formulas}{formulae for variables}

\item{pars}{list of parameters}

\item{family}{list of numeric vectors in same format as \code{formulas}}

\item{lags}{numeric not currently used}

\item{dat}{data frame...}

\item{control}{list of control parameters}
}
\description{
Simulate from a discrete time survival model
}
\details{
variables in family take their usual format, though \code{Y} is
assumed to be binary if unspecified.
}
