% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_vinecop.R
\name{sim_vinecop}
\alias{sim_vinecop}
\title{Simulate from vine copula}
\usage{
sim_vinecop(dat, family, par, par2 = NULL, model_matrix)
}
\arguments{
\item{dat}{data frame to be filled in}

\item{family}{family to simulate from}

\item{par}{matrix of parameters}

\item{par2}{extra parameters for t-copula}

\item{model_matrix}{design matrix for covariates}
}
\value{
A data frame of the same dimensions as \code{dat}.
}
\description{
Simulate from vine copula
}
