% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\name{sim_X}
\alias{sim_X}
\title{Simulate initial X values}
\usage{
sim_X(n, fam_x, theta)
}
\arguments{
\item{n}{number of observations}

\item{fam_x}{number for distribution family}

\item{theta}{parameters for model}
}
\value{
A list with two entries: \code{x} a vector of the simulated
values, and \code{qden}, which contains a function that evaluates to the
density of the distribution used to generate those values.
}
\description{
Simulate initial X values
}
\details{
Returns a list that includes a data frame containing a column
\code{x}, as well as the density that was used to generate it.  Possible
families are Gaussian (=1), t (=2), Exponential (=3), beta (=4)
Bernoulli/categorical (=5).

For the exponential distribution, \code{theta} is the mean.
Beta can take one or two parameters, and if there is only
one it is just repeated.
}
