% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_inversion.R
\name{sim_Y}
\alias{sim_Y}
\title{Generate outcome data by inversion}
\usage{
sim_Y(n, formulas, family, pars, formY, famY, parsY, linkY, qZ, vars, dat)
}
\arguments{
\item{n}{number of samples}

\item{formulas}{list of formulae for copula parameters}

\item{family}{vector of integers for distributions of variables}

\item{pars}{vector of parameters for copula}

\item{formY}{formulae for response variable}

\item{famY}{family for response variable}

\item{parsY}{regression parameters for response variable}

\item{linkY}{link function for response variable}

\item{qZ}{quantiles of covariate distribution}

\item{vars}{character vector of variable names}

\item{dat}{data frame containing variables in \code{formulas} and \code{formY}}
}
\description{
Generate outcome data by inversion
}
