% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{theta_to_p_ord}
\alias{theta_to_p_ord}
\alias{theta_to_p_cat}
\title{Transform categorical or ordinal parameters into probabilities}
\usage{
theta_to_p_ord(theta)

theta_to_p_cat(theta)
}
\arguments{
\item{theta}{provided log-linear parameters}
}
\description{
Transform categorical or ordinal parameters into probabilities
}
\details{
Returns the probabilities implied by given log-linear parameters.
}
\section{Functions}{
\itemize{
\item \code{theta_to_p_ord()}: for ordinal variables

}}
