% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intialize.R
\name{tidy_formulas}
\alias{tidy_formulas}
\title{Tidy up formulae}
\usage{
tidy_formulas(formulas, kwd, prefix = "V")
}
\arguments{
\item{formulas}{list of model formulae, for Y, for the Z variables, and
finally for the copula}

\item{kwd}{string used to denote copula}

\item{prefix}{string to begin each new variable name}
}
\description{
This function ensures that all formulae have a left hand side, by giving
them names of the form \code{Vn} where \code{n} is some positive integer.
}
