% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulas.R
\name{cVCopula}
\alias{cVCopula}
\title{Faster Vectorized conditional copula}
\usage{
cVCopula(U, copula, param, par2 = NULL, inverse = FALSE, cdf = FALSE)
}
\arguments{
\item{U}{matrix of quantiles}

\item{copula}{family of copula to use}

\item{param}{vector of parameters}

\item{par2}{degrees of freedom for t-copula}

\item{inverse}{should inverse CDF be returned?}

\item{cdf}{should we evaluate the CDF copula, not the conditional?}
}
\description{
This is a faster version of \code{cCopula} where we don't build redundant
copulas if \code{unique(eta)} is small.
}
\details{
Should have either \code{length(param) = 1} or \code{nrow(U) = length(param)}.
}
