% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copulas.R
\name{cVCopula}
\alias{cVCopula}
\title{A faster Vectorized conditional copula function
where we don't build redudendant copulas if unique(eta) is small}
\usage{
cVCopula(U, copula, param, par2 = NULL, inverse = FALSE, cdf = FALSE)
}
\arguments{
\item{U}{matrix of quantiles}

\item{copula}{family of copula to use}

\item{param}{vector of parameters}

\item{par2}{Degrees of freedom for t-copula}

\item{inverse}{should inverse CDF be returned?}

\item{cdf}{should we evaluate the cdf copula, not the conditional?}
}
\description{
A faster Vectorized conditional copula function
where we don't build redudendant copulas if unique(eta) is small
}
\details{
Should have \code{nrow(U) = length(param)}.
}
