% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\docType{data}
\name{familyVals}
\alias{familyVals}
\alias{copulaVals}
\title{Numbers for parametric families}
\format{
An object of class \code{data.frame} with 7 rows and 3 columns.

An object of class \code{data.frame} with 7 rows and 2 columns.
}
\usage{
familyVals

copulaVals
}
\description{
Each function returns a data frame containing
\itemize{
\item \code{val}: an integer
\item \code{family}: a vector giving the associated parametric family for that integer.
}
}
\section{Functions}{
\itemize{
\item \code{copulaVals}: Values for copula families

}}
\keyword{datasets}
