% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\docType{data}
\name{familyVals}
\alias{familyVals}
\alias{copula_vals}
\title{Numbers for parametric families}
\format{
An object of class \code{data.frame} with 9 rows and 2 columns.

An object of class \code{data.frame} with 7 rows and 2 columns.
}
\usage{
familyVals

copula_vals
}
\description{
Each function returns a data frame containing
\itemize{
\item \code{val}: an integer
\item \code{family}: a vector giving the associated parametric family for that integer.
}
}
\section{Functions}{
\itemize{
\item \code{copula_vals}: Values for copula families

}}
\keyword{datasets}
