% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_causl}
\alias{fit_causl}
\alias{fitCausal}
\title{Fit multivariate copula regression model}
\usage{
fit_causl(
  dat,
  formulas = list(y ~ x, z ~ 1, ~x),
  family = rep(1, length(formulas)),
  link,
  cop_pars,
  use_cpp = TRUE,
  control = list(),
  other_pars = list(),
  method = "optim"
)

fitCausal(
  dat,
  formulas = list(y ~ x, z ~ 1, ~x),
  family = rep(1, length(formulas)),
  link,
  par2,
  sandwich = TRUE,
  use_cpp = TRUE,
  control = list()
)
}
\arguments{
\item{dat}{data frame of observations}

\item{formulas}{list of model formulae, for outcome, covariates, and
finally for the copula}

\item{family}{families for variables and copula as above. Should be the
same length as \code{formulas}, and in same order}

\item{link}{link functions for each variable}

\item{cop_pars}{additional parameters for copula if required}

\item{use_cpp}{logical: should C++ routines be used?}

\item{control}{list of parameters to be passed to \code{optim}}

\item{other_pars}{list of other parameters to use (e.g. degrees of freedom for a t-distribution)}

\item{method}{currently only \code{"optim"} is supported}

\item{par2}{former name for \code{cop_pars} argument}

\item{sandwich}{logical: should sandwich standard errors be returned?}
}
\value{
Returns a list of class \code{cop_fit}.
}
\description{
Fit multivariate copula regression model
}
\details{
\code{formulas} is list of three or more formulae giving predictors of
y-margin, z-margin(s) and interaction parameters.  Fit is by maximum
likelihood.

\code{family} takes numeric values of the following forms:\tabular{rl}{
   val \tab family \cr
   0 \tab binomial \cr
   1 \tab gaussian \cr
   2 \tab t \cr
   3 \tab Gamma \cr
   4 \tab beta \cr
   5 \tab binomial \cr
   6 \tab lognormal \cr
   10 \tab categorical \cr
   11 \tab ordinal \cr
}


\code{control} has the same arguments as the argument in \code{optim}, as well
as \code{sandwich}, a logical indicating if sandwich estimates of standard errors
should be computed, \code{newton}, a logical which controls whether Newton iterates should be
performed at the end, and \code{cop} which can edit the restricted variable name
for the left-hand side of formulae.
Useful for altering are \code{trace} (1 shows steps of optimization) and
\code{maxit} for the number of steps.

\code{cop_pars} is currently just a numeric scalar, generally representing the
degrees of freedom to use in a t-copula.

The list \code{other_pars} should be named with the relevant variables, and
each entry should be a named list containing the relevant parameters.  As
an example, for a t-copula, the required parameter is the degrees of
freedom, \code{df}.

\strong{Warning} By default, none of the variables should be called \code{cop}, as
this is reserved for the copula.  The reserved word can be changed using
the argument \code{cop} within control.
}
\section{Functions}{
\itemize{
\item \code{fitCausal()}: old name

}}
