% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{gen_dummy_dat}
\alias{gen_dummy_dat}
\title{Generate a dummy dataset}
\usage{
gen_dummy_dat(family, pars, dat, LHSs, dims)
}
\arguments{
\item{family}{families for variables and copula}

\item{pars}{list of lists of parameters}

\item{dat}{optional data frame of covariates}

\item{LHSs}{left-hand sides from \code{formulas}}

\item{dims}{number of variables in each class}
}
\description{
Create a dummy dataset for the purpose of checking coefficient numbers
}
