% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{get_family}
\alias{get_family}
\alias{gaussian_causl_fam}
\alias{t_causl_fam}
\alias{Gamma_causl_fam}
\alias{binomial_causl_fam}
\alias{beta_causl_fam}
\alias{categorical_causl_fam}
\alias{ordinal_causl_fam}
\title{Return causl_fam function from integer index}
\usage{
get_family(val)

gaussian_causl_fam(link)

t_causl_fam(link)

Gamma_causl_fam(link)

binomial_causl_fam(link)

beta_causl_fam(link)

categorical_causl_fam(link)

ordinal_causl_fam(link)
}
\arguments{
\item{val}{integer corresponding to distributional family}

\item{link}{link function}
}
\description{
Return causl_fam function from integer index
}
\details{
The functions \code{gaussian_causl_fam()} etc. represent the functions that are
returned by \code{get_family()}.
}
\section{Functions}{
\itemize{
\item \code{gaussian_causl_fam()}: Gaussian distribution family

\item \code{t_causl_fam()}: Student's t distribution family

\item \code{Gamma_causl_fam()}: Gamma distribution family

\item \code{binomial_causl_fam()}: binomial distribution family

\item \code{beta_causl_fam()}: beta distribution family

\item \code{categorical_causl_fam()}: multinomial/categorical distribution family

\item \code{ordinal_causl_fam()}: ordinal categorical distribution family

}}
\seealso{
\link{family_vals}
}
