% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\name{sim_glm}
\alias{sim_glm}
\alias{glm_sim}
\title{Simulate from a GLM}
\usage{
sim_glm(family, eta, phi, par2, link)

glm_sim(family, eta, phi, par2, link)
}
\arguments{
\item{family}{vector of distribution families}

\item{eta}{list (or matrix) of linear forms}

\item{phi}{vector of dispersion coefficients}

\item{par2}{vector of degrees of freedom}

\item{link}{link functions for GLMs}
}
\value{
a numeric vector of weights
}
\description{
Simulate values from some generalized linear models
}
\section{Functions}{
\itemize{
\item \code{sim_glm()}: Old name

}}
