% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\name{glm_sim}
\alias{glm_sim}
\title{Simulate from a GLM}
\usage{
glm_sim(family, eta, phi, other_pars, link)
}
\arguments{
\item{family}{vector of distribution families}

\item{eta}{list (or matrix) of linear forms}

\item{phi}{vector of dispersion coefficients}

\item{other_pars}{list of other parameters for specified family}

\item{link}{link functions for GLMs}
}
\value{
a numeric vector of weights
}
\description{
Simulate values from some generalized linear models
}
