% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{lhs}
\alias{lhs}
\title{Get LHSs of list of formulae}
\usage{
lhs(formulas)
}
\arguments{
\item{formulas}{list of formulae}
}
\description{
Get LHSs of list of formulae
}
\details{
Returns character vector containing left-hand sides of a list of
formulae.
}
