% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{lhs}
\alias{lhs}
\alias{lhs<-}
\title{Get LHSs of list of formulae}
\usage{
lhs(formulas)

lhs(formulas) <- value
}
\arguments{
\item{formulas}{list of formulae}

\item{value}{charachter vector to assign}
}
\description{
Get LHSs of list of formulae
}
\details{
Returns character vector containing left-hand sides of a list of
formulae.
}
\section{Functions}{
\itemize{
\item \code{lhs(formulas) <- value}: Assign LHSs to list of formulas

}}
