% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\name{link_apply}
\alias{link_apply}
\title{Helper Function to apply Link Function}
\usage{
link_apply(eta, link, family_nm)
}
\arguments{
\item{eta}{inner product of X and beta}

\item{link}{link function}

\item{family_nm}{family of distributions to use}
}
\value{
link transformed inner product eta
}
\description{
Helper Function to apply Link Function
}
