% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{link_setup}
\alias{link_setup}
\title{Set up link functions}
\usage{
link_setup(link, family, vars)
}
\arguments{
\item{link}{the input given to causalSamp()}

\item{family}{the list of families for Z,X and Y variables}

\item{vars}{a list of vectors of variable names with the same structure as \code{family}}
}
\description{
Set up link functions
}
