% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_vinecop.R
\name{links_cop}
\alias{links_cop}
\title{List of link functions for copuulas}
\usage{
links_cop(family)
}
\arguments{
\item{family}{either the name of the family or its integer representative}
}
\description{
List of link functions for copuulas
}
\details{
This function returns the default link function for each possible copula.
These are given in the table:

|value|family     |link                  |  link name |
|-----:|----------|----------------------|---|
|1    |gaussian  |logit((1+rho)/2)      | logit2 |
|2    |t         |logit((1+rho)/2)      |logit2 |
|3    |Clayton   |log(1+theta)      | log1p |
|4    |Gumbel    |log(theta - 1)      | log1m |
|5    |Frank     |logit(1 + theta)      | log1p |
|6    |Joe         |identity      | identity |
|11    |FGM         |logit((1+rho)/2)      | logit2 |
}
