% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{process_inputs}
\alias{process_inputs}
\title{Process formulas, families and parameters}
\usage{
process_inputs(formulas, pars, family, link, dat, kwd, ordering = FALSE)
}
\arguments{
\item{formulas}{list of lists of formulas}

\item{pars}{list of lists of parameters}

\item{family}{families for variables and copula}

\item{link}{list of link functions}

\item{dat}{optional data frame of covariates}

\item{kwd}{keyword for copula}

\item{ordering}{logical: should an ordering of variables be computed?}
}
\description{
Process formulas, families and parameters
}
