% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{process_inputs}
\alias{process_inputs}
\alias{process_formulas}
\alias{process_family}
\title{Process formulas, families and parameters}
\usage{
process_inputs(formulas, pars, family, link, dat, kwd, method = "inversion")

process_formulas(formulas)

process_family(family, dims)
}
\arguments{
\item{formulas}{list of lists of formulas}

\item{pars}{list of lists of parameters}

\item{family}{families for variables and copula}

\item{link}{list of link functions}

\item{dat}{optional data frame of covariates}

\item{kwd}{keyword for copula}

\item{method}{either \code{"inversion"} (the default) or \code{"rejection"}}

\item{dims}{number of variables in each class}
}
\description{
Process formulas, families and parameters
}
\section{Functions}{
\itemize{
\item \code{process_formulas()}: Process input for family variables

\item \code{process_family()}: Process input for family variables

}}
