% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{process_inputs}
\alias{process_inputs}
\alias{process_formulas}
\alias{process_family}
\title{Process formulas, families and parameters}
\usage{
process_inputs(formulas, pars, family, link, dat, kwd, method = "inversion")

process_formulas(formulas, len = 4)

process_family(family, dims, func_return = get_family)
}
\arguments{
\item{formulas}{list of lists of formulas}

\item{pars}{list of lists of parameters}

\item{family}{families for variables and copula}

\item{link}{list of link functions}

\item{dat}{optional data frame of covariates}

\item{kwd}{keyword for copula}

\item{method}{either \code{"inversion"} (the default) or \code{"rejection"}}

\item{dims}{number of variables in each class}

\item{func_return}{function to use to process character arguments}
}
\description{
Process formulas, families and parameters
}
\details{
Function that processes and checks the validity of the main arguments
used for simulating data.

For \code{causl} we use the \code{get_family()} function to process character based
arguments, but we allow for other functions to be used in packages that
build on this one.
}
\section{Functions}{
\itemize{
\item \code{process_formulas()}: Process input for family variables

\item \code{process_family()}: Process input for family variables

}}
