% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_funcs.R
\name{rescale_cop}
\alias{rescale_cop}
\alias{rescaleCop}
\title{Rescale quantiles to conditional copula}
\usage{
rescale_cop(U, X, par, family = 1, df, inv = TRUE, cdf = FALSE)

rescaleCop(U, X, beta, family = 1, df)
}
\arguments{
\item{U}{matrix of quantiles}

\item{X}{model matrix of covariates}

\item{family}{variety of copula to use}

\item{beta}{list of parameters (see details)}
}
\value{
vector of rescaled quantiles
}
\description{
Rescale quantiles to conditional copula
}
\details{
The variable to be transformed must be in the final column of
\code{U}, with variables being conditioned upon in the earlier columns.

\code{family} can be 1 for Gaussian, 2 for t, 3 for Clayton, 4 for
Gumbel, 5 for Frank, 6 for Joe and 11 for FGM copulas. Gamma distributed,
beta distributed or discrete respectively. \code{beta} should contain regression
parameters, \code{df} must be specified if \code{family=2}. \code{U} should have the
same length as \code{X} has rows, and \code{X}
should have the same number of columns as the length of \code{pars$beta}.
}
\section{Functions}{
\itemize{
\item \code{rescaleCop()}: Old name, now deprecated

}}
