% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_optim.R
\name{sandwich_errors}
\alias{sandwich_errors}
\title{Compute robust standard errors}
\usage{
sandwich_errors(func, pars, method = "simple", trace = 0L, other_args)
}
\arguments{
\item{func}{function to be optimized}

\item{pars}{point at which function should be optimized}

\item{method}{taken from \code{grad()}; defaults to \code{"simple"} for computational reasons}

\item{trace}{provides more details if > 0.}

\item{other_args}{other arguments}
}
\description{
Compute robust standard errors
}
