% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_inversion.R, R/sim_multi.R,
%   R/sim_rejection.R
\name{sim_inversion}
\alias{sim_inversion}
\alias{sim_multi}
\alias{sim_rejection}
\title{Simulate for single time-step}
\usage{
sim_inversion(out, proc_inputs)

sim_multi(out, proc_inputs)

sim_rejection(out, proc_inputs, careful)
}
\arguments{
\item{out}{data frame for output}

\item{proc_inputs}{output of \code{process_inputs()}}

\item{careful}{should full, slower method be used?}
}
\description{
Simulate for single time-step
}
\details{
\code{sim_inversion} and \code{sim_rejection} correspond to
performing the sampling by inversion or using rejection sampling.

\code{sim_multi} first simulates from the copula then transforms to the
correct margins in the correct causal ordering
}
\section{Functions}{
\itemize{
\item \code{sim_multi()}: simulation with multivariate copula

\item \code{sim_rejection()}: Rejection sampling code

}}
