% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nll.R
\name{univarDens}
\alias{univarDens}
\title{Get univariate densities and uniform order statistics}
\usage{
univarDens(x, eta, phi, df, family = 1, link)
}
\arguments{
\item{x}{vector of observations}

\item{eta, phi}{linear component and dispersion parameters}

\item{df}{degrees of freedom (only for t-distribution)}

\item{family}{numeric indicator of family}

\item{link}{link function}
}
\value{
A list with entries being the numeric vectors \code{u} (the
quantiles of the input values) and \code{ld} (the log density of each
observation).
}
\description{
Get univariate densities and uniform order statistics
}
\details{
\code{fam} follows the usual numeric pattern: 1=normal,
2=t-distribution and 3=Gamma with a log-link.
}
